MaterialId u8 → u16: lifted 256-material cap, modders can now add up to 65535
Refactored the core MaterialId type from u8 to u16 across the workspace.
This lifts the 256-material ceiling that was about to bite modders (89/256 used).

Mechanical changes:
- cf-terrain::chunked: MaterialId = u16; Chunk::pixels is Vec<u16>
- cf-material::lib: MaterialId = u16 alias
- ReactionLookup: replaced 256x256 dense table (impossible at 65536x65536)
  with HashMap<(MaterialId, MaterialId), Vec<u32>>; same O(1) avg lookup
- primary_reactive_bitmap: replaced [bool; 256] with Vec<bool> sized to
  max_material_id+1 so it grows with the registry
- checksum_bytes: bumped to v2; pixels hashed as LE bytes per material
- inspect_material API: id parameter widened to u16
- material_distribution map: now BTreeMap<u16, u64>
- Chunk pixel buffer: doubled in size (1 byte → 2 bytes per pixel)
  - 1024x1024 world: 1 MB → 2 MB (still trivial)
- u16_slice_to_bytes helper in engine.rs for hex encoding of pixel arrays
- Inspect-material schema v1 file regenerated with uint16 format
- 25 .rs files touched; all type-driven by the compiler

Tests: 4293 passing (no regressions).

+++ game/crates/cf-app/src/main.rs
@@ -1122,12 +1122,12 @@ fn pump_recorder_events_into_render_effects(
                     .payload
                     .get("spawn_material_id")
                     .and_then(|v| v.as_u64())
-                    .and_then(|n| u8::try_from(n).ok())
+                    .and_then(|n| u16::try_from(n).ok())
                     .or_else(|| {
                         ev.payload
                             .get("source_material_id")
                             .and_then(|v| v.as_u64())
-                            .and_then(|n| u8::try_from(n).ok())
+                            .and_then(|n| u16::try_from(n).ok())
                     })
                     .unwrap_or(cf_terrain::MATERIAL_LOOSE_FILL);
                 debris_queue.pending.push_back(DebrisSpawnRequest {
